Add the state_based mode for DB pollers.
