+++ lib/deimos/utils/db_poller/time_based.rb
@@ -1,0 +1,79 @@
+require 'deimos/utils/db_poller/base'
+
+module Deimos
+  module Utils
+    module DbPoller
+      class TimeBased < Base
+
+        # :nodoc:
+        def create_poll_info
+          new_time = @config.start_from_beginning ? Time.new(0) : Time.zone.now
+          Deimos::PollInfo.create!(producer: @config.producer_class,
+                                           last_sent: new_time,
+                                           last_sent_id: 0)
+        end
+
+        # @param batch [Array<ActiveRecord::Base>]
+        # @param status [Deimos::Utils::DbPoller::PollStatus]
+        def process_and_touch_info(batch, status)
+          process_batch_with_span(batch, status)
+          self.touch_info(batch)
+        end
+
+        # Send messages for updated data.
+        # @return [void]
+        def process_updates
+          time_from = @config.full_table ? Time.new(0) : @info.last_sent.in_time_zone
+          time_to = Time.zone.now - @config.delay_time
+          Deimos.config.logger.info("Polling #{@producer.topic} from #{time_from} to #{time_to}")
+          status = PollStatus.new(0, 0, 0)
+
+          # poll_query gets all the relevant data from the database, as defined
+          # by the producer itself.
+          loop do
+            Deimos.config.logger.debug("Polling #{@producer.topic}, batch #{status.current_batch}")
+            batch = fetch_results(time_from, time_to).to_a
+            break if batch.empty?
+
+            process_and_touch_info(batch, status)
+            time_from = last_updated(batch.last)
+          end
+          Deimos.config.logger.info("Poll #{@producer.topic} complete at #{time_to} (#{status.report})")
+        end
+
+        # @param time_from [ActiveSupport::TimeWithZone]
+        # @param time_to [ActiveSupport::TimeWithZone]
+        # @return [ActiveRecord::Relation]
+        def fetch_results(time_from, time_to)
+          id = @producer.config[:record_class].primary_key
+          quoted_timestamp = ActiveRecord::Base.connection.quote_column_name(@config.timestamp_column)
+          quoted_id = ActiveRecord::Base.connection.quote_column_name(id)
+          @producer.poll_query(time_from: time_from,
+                               time_to: time_to,
+                               column_name: @config.timestamp_column,
+                               min_id: @info.last_sent_id).
+            limit(BATCH_SIZE).
+            order("#{quoted_timestamp}, #{quoted_id}")
+        end
+
+        # @param record [ActiveRecord::Base]
+        # @return [ActiveSupport::TimeWithZone]
+        def last_updated(record)
+          record.public_send(@config.timestamp_column)
+        end
+
+        # @param batch [Array<ActiveRecord::Base>]
+        # @return [void]
+        def touch_info(batch)
+          record = batch.last
+          id_method = record.class.primary_key
+          last_id = record.public_send(id_method)
+          last_updated_at = last_updated(record)
+          @info.attributes = { last_sent: last_updated_at, last_sent_id: last_id }
+          @info.save!
+        end
+
+      end
+    end
+  end
+end
