+++ rbs_collection.lock.yaml
@@ -11,7 +11,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: actionview
@@ -19,7 +19,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: activejob
@@ -27,7 +27,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: activemodel
@@ -35,7 +35,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: activerecord
@@ -43,7 +43,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: activestorage
@@ -51,7 +51,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: activesupport
@@ -59,7 +59,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: ast
@@ -67,19 +67,23 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
-- name: deimos-ruby
-  version: 1.16.4
+- name: dogstatsd-ruby
+  version: '5.5'
   source:
-    type: rubygems
+    type: git
+    name: ruby/gem_rbs_collection
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
+    remote: https://github.com/ruby/gem_rbs_collection.git
+    repo_dir: gems
 - name: i18n
   version: '1.10'
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: listen
@@ -87,7 +91,15 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
+    remote: https://github.com/ruby/gem_rbs_collection.git
+    repo_dir: gems
+- name: mail
+  version: '2.8'
+  source:
+    type: git
+    name: ruby/gem_rbs_collection
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: minitest
@@ -99,7 +111,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: parallel
@@ -107,7 +119,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: phobos
@@ -115,11 +127,11 @@ gems:
   source:
     type: rubygems
 - name: rack
-  version: 2.2.2
+  version: '2.2'
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: railties
@@ -127,7 +139,7 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
 - name: rainbow
@@ -135,9 +147,13 @@ gems:
   source:
     type: git
     name: ruby/gem_rbs_collection
-    revision: 8e2fd75fc0e2bcbda9023e8c936fe5ba654b1436
+    revision: ec15faa1e16b3f4596eb80eabc902bd2589028cf
     remote: https://github.com/ruby/gem_rbs_collection.git
     repo_dir: gems
+- name: steep
+  version: 1.1.1
+  source:
+    type: rubygems
 - name: tempfile
   version: '0'
   source:
@@ -167,6 +183,14 @@ gems:
   source:
     type: stdlib
 - name: pathname
+  version: '0'
+  source:
+    type: stdlib
+- name: cgi
+  version: '0'
+  source:
+    type: stdlib
+- name: uri
   version: '0'
   source:
     type: stdlib
